+++ eyes.sdk.core/src/main/java/com/applitools/eyes/visualgrid/model/HashObject.java
@@ -1,0 +1,19 @@
+package com.applitools.eyes.visualgrid.model;
+
+public class HashObject {
+    private final String hashFormat;
+    private final String hash;
+
+    public HashObject(String hashFormat, String hash) {
+        this.hashFormat = hashFormat;
+        this.hash = hash;
+    }
+
+    public String getHashFormat() {
+        return hashFormat;
+    }
+
+    public String getHash() {
+        return hash;
+    }
+}