+++ eyes.sdk.core/src/main/java/com/applitools/eyes/visualgrid/model/RenderingTask.java
@@ -2,11 +2,12 @@ package com.applitools.eyes.visualgrid.model;
 
 import com.applitools.eyes.EyesException;
 import com.applitools.eyes.Logger;
+import com.applitools.eyes.SyncTaskListener;
 import com.applitools.eyes.TaskListener;
-import com.applitools.eyes.UserAgent;
 import com.applitools.eyes.visualgrid.services.IEyesConnector;
 import com.applitools.eyes.visualgrid.services.VisualGridRunner;
 import com.applitools.eyes.visualgrid.services.VisualGridTask;
+import com.applitools.utils.EyesSyncObject;
 import com.applitools.utils.GeneralUtils;
 
 import java.util.*;
@@ -15,6 +16,7 @@ import java.util.concurrent.Phaser;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicReference;
 
 public class RenderingTask implements Callable<RenderStatusResults> {
 
@@ -27,6 +29,7 @@ public class RenderingTask implements Callable<RenderStatusResults> {
     private final List<VisualGridTask> checkTasks = new ArrayList<>();
     final Map<String, RGridResource> fetchedCacheMap;
     final Map<String, RGridResource> putResourceCache;
+    final Set<String> checkResourceCache;
     private final Logger logger;
     private final AtomicBoolean isForcePutNeeded;
     private final Timer timer = new Timer("VG_StopWatch", true);
@@ -68,6 +71,7 @@ public class RenderingTask implements Callable<RenderStatusResults> {
         this.checkTasks.add(checkTask);
         this.fetchedCacheMap = runner.getCachedResources();
         this.putResourceCache = runner.getPutResourceCache();
+        this.checkResourceCache = runner.getCheckResourceCache();
         this.logger = runner.getLogger();
         this.listener = listener;
         String renderingGridForcePut = GeneralUtils.getEnvString("APPLITOOLS_RENDERING_GRID_FORCE_PUT");
@@ -84,7 +88,20 @@ public class RenderingTask implements Callable<RenderStatusResults> {
         try {
             logger.verbose("enter");
 
-            logger.verbose("step 1");
+            logger.verbose("Uploading missing resources");
+            resourcesPhaser = new Phaser();
+            Map<String, RGridResource> missingResources = checkResourcesStatus();
+            uploadResources("NONE", missingResources.keySet(), missingResources, false);
+            try {
+                if (resourcesPhaser.getRegisteredParties() > 0) {
+                    resourcesPhaser.awaitAdvanceInterruptibly(0, 30, TimeUnit.SECONDS);
+                }
+            } catch (InterruptedException | TimeoutException e) {
+                GeneralUtils.logExceptionStackTrace(logger, e);
+                resourcesPhaser.forceTermination();
+            }
+
+            logger.verbose("Start rendering");
             boolean stillRendering;
             long elapsedTimeStart = System.currentTimeMillis();
             boolean isForcePutAlreadyDone = false;
@@ -106,7 +123,7 @@ public class RenderingTask implements Callable<RenderStatusResults> {
                     }
                 }
 
-                logger.verbose("step 2.1");
+                logger.verbose("Validation render result");
                 if (runningRenders == null || runningRenders.size() == 0) {
                     setRenderErrorToTasks();
                     throw new EyesException("Invalid response for render request");
@@ -118,7 +135,6 @@ public class RenderingTask implements Callable<RenderStatusResults> {
                     logger.verbose(String.format("RunningRender: %s", runningRenders.get(i)));
                 }
 
-                logger.verbose("step 2.2");
                 boolean shouldUploadResources = shouldUploadResources(runningRenders);
                 double elapsedTime = ((double) System.currentTimeMillis() - elapsedTimeStart) / 1000;
                 stillRendering = (shouldUploadResources && elapsedTime < FETCH_TIMEOUT_SECONDS);
@@ -126,6 +142,7 @@ public class RenderingTask implements Callable<RenderStatusResults> {
                     continue;
                 }
 
+                logger.verbose("Uploading missing resources");
                 boolean forcePut = isForcePutNeeded.get() && !isForcePutAlreadyDone;
                 uploadResources(runningRenders, forcePut);
                 if (forcePut) {
@@ -139,12 +156,9 @@ public class RenderingTask implements Callable<RenderStatusResults> {
                     GeneralUtils.logExceptionStackTrace(logger, e);
                     resourcesPhaser.forceTermination();
                 }
-
-                logger.verbose("step 2.3");
-
             } while (stillRendering);
 
-            logger.verbose("step 3");
+            logger.verbose("Poll rendering status");
             Map<RunningRender, RenderRequest> mapping = mapRequestToRunningRender(runningRenders);
             pollRenderingStatus(mapping);
         } catch (Throwable e) {
@@ -157,6 +171,57 @@ public class RenderingTask implements Callable<RenderStatusResults> {
         logger.verbose("Finished rendering task - exit");
 
         return null;
+    }
+
+    /**
+     * Checks with the server what resources are missing.
+     * @return All the missing resources to upload.
+     */
+    Map<String, RGridResource> checkResourcesStatus() {
+        List<HashObject> hashesToCheck = new ArrayList<>();
+        Map<String, String> hashToResourceUrl = new HashMap<>();
+        for (Map.Entry<String, RGridResource> pair : fetchedCacheMap.entrySet()) {
+            String hash = pair.getValue().getSha256();
+            String hashFormat = pair.getValue().getHashFormat();
+            synchronized (checkResourceCache) {
+                if (!checkResourceCache.contains(hash)) {
+                    hashesToCheck.add(new HashObject(hashFormat, hash));
+                    hashToResourceUrl.put(hash, pair.getKey());
+                    checkResourceCache.add(hash);
+                }
+            }
+        }
+
+        if (hashesToCheck.isEmpty()) {
+            return new HashMap<>();
+        }
+
+        AtomicReference<EyesSyncObject> lock = new AtomicReference<>(new EyesSyncObject(logger, "checkResourceStatus"));
+        AtomicReference<Boolean[]> reference = new AtomicReference<>();
+        SyncTaskListener<Boolean[]> listener = new SyncTaskListener<>(lock, reference);
+        eyesConnector.checkResourceStatus(listener, null, hashesToCheck.toArray(new HashObject[0]));
+        synchronized (lock.get()) {
+            try {
+                lock.get().waitForNotify();
+            } catch (InterruptedException ignored) {}
+        }
+
+        Boolean[] result = reference.get();
+        if (result == null) {
+            return new HashMap<>();
+        }
+
+        Map<String, RGridResource> missingResources = new HashMap<>();
+        for (int i = 0; i < result.length; i++) {
+            if (result[i] != null && result[i]) {
+                continue;
+            }
+
+            String resourceUrl = hashToResourceUrl.get(hashesToCheck.get(i).getHash());
+            missingResources.put(resourceUrl, fetchedCacheMap.get(resourceUrl));
+        }
+
+        return missingResources;
     }
 
     void uploadResources(List<RunningRender> runningRenders, boolean forcePut) {
@@ -172,7 +237,7 @@ public class RenderingTask implements Callable<RenderStatusResults> {
             if (runningRender.isNeedMoreDom() || forcePut) {
                 try {
                     resourcesPhaser.register();
-                    this.eyesConnector.renderPutResource(runningRender, dom.asResource(), putListener);
+                    this.eyesConnector.renderPutResource(runningRender.getRenderId(), dom.asResource(), putListener);
                 } catch (Throwable e) {
                     GeneralUtils.logExceptionStackTrace(logger, e);
                 }
@@ -181,36 +246,40 @@ public class RenderingTask implements Callable<RenderStatusResults> {
             // Uploading missing resources
             logger.verbose("creating PutFutures for " + runningRenders.size() + " runningRenders");
             Collection<String> urls = forcePut ? resources.keySet() : runningRender.getNeedMoreResources();
-            for (String url : urls) {
-                if (putResourceCache.containsKey(url) && ! forcePut) {
-                    continue;
-                }
+            uploadResources(runningRender.getRenderId(), urls, resources, forcePut);
+        }
+        logger.verbose("exit");
+    }
 
-                RGridResource resource;
-                if (!fetchedCacheMap.containsKey(url)) {
-                    logger.verbose(String.format("Resource %s requested but never downloaded (maybe a Frame)", url));
-                    resource = resources.get(url);
-                } else {
-                    resource = fetchedCacheMap.get(url);
-                }
+    private void uploadResources(String renderId, Collection<String> urls, Map<String, RGridResource> resources, boolean forcePut) {
+        for (String url : urls) {
+            if (putResourceCache.containsKey(url) && ! forcePut) {
+                continue;
+            }
 
-                if (resource == null) {
-                    logger.log(String.format("Illegal state: resource is null for url %s", url));
-                    continue;
-                }
+            RGridResource resource;
+            if (!fetchedCacheMap.containsKey(url)) {
+                logger.verbose(String.format("Resource %s requested but never downloaded (maybe a Frame)", url));
+                resource = resources.get(url);
+            } else {
+                resource = fetchedCacheMap.get(url);
+            }
 
-                logger.verbose("resource(" + resource.getUrl() + ") hash : " + resource.getSha256());
-                resourcesPhaser.register();
-                this.eyesConnector.renderPutResource(runningRender, resource, putListener);
-                String contentType = resource.getContentType();
-                synchronized (putResourceCache) {
-                    if (!putResourceCache.containsKey(url) && (contentType != null && !contentType.equalsIgnoreCase(RGridDom.CONTENT_TYPE))) {
-                        putResourceCache.put(url, resource);
-                    }
+            if (resource == null) {
+                logger.log(String.format("Illegal state: resource is null for url %s", url));
+                continue;
+            }
+
+            logger.verbose("resource(" + resource.getUrl() + ") hash : " + resource.getSha256());
+            resourcesPhaser.register();
+            this.eyesConnector.renderPutResource(renderId, resource, putListener);
+            String contentType = resource.getContentType();
+            synchronized (putResourceCache) {
+                if (!putResourceCache.containsKey(url) && (contentType != null && !contentType.equalsIgnoreCase(RGridDom.CONTENT_TYPE))) {
+                    putResourceCache.put(url, resource);
                 }
             }
         }
-        logger.verbose("exit");
     }
 
     private void setRenderErrorToTasks() {